Introduction of additional props to the reusable Input component.
